tests/wpt/tests/html/canvas/element/path-objects/2d.path.open.arc.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.path.open.arc</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.open.arc</h1>
<p class="desc">line caps should be drawn on open arcs</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="300" height="300"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("line caps should be drawn on open arcs");
_addTest(function(canvas, ctx) {

  ctx.beginPath();
  ctx.strokeStyle = 'red';
  ctx.lineWidth = 20;
  ctx.lineCap = 'square';
  ctx.arc(150, 150, 50, 0, 2 * Math.PI);
  ctx.stroke();
  // The pixel (209, 148) is part of the line cap, but not part of the circle.
  // Linecap is drawn with 360 degree arc, so the pixel is red.
  _assertPixelApprox(canvas, 209,148, 255,0,0,255, 1);

});
</script>
